requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
realesrgan>=0.3.0
gfpgan>=1.3.8
basicsr>=1.4.2
numpy>=1.21.0
